requirements.txt
pysimplegui
numpy
Pillow